% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_Dynameta.R
\name{launch_Dynameta}
\alias{launch_Dynameta}
\title{Launch Dynameta}
\usage{
launch_Dynameta()
}
\value{
Shiny application object
}
\description{
Launches the Dynameta shiny app
}
\examples{
\dontrun{launch_Dynameta()}

}
