% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynameta_Server.R
\name{server}
\alias{server}
\title{Dynameta shiny app server function}
\usage{
server(input, output)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}
}
\description{
Script that defines the server function of Dynameta shiny app
}
