% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynameta_UI.R
\docType{data}
\name{ui}
\alias{ui}
\title{Dynameta shiny app ui script}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 5.
}
\usage{
ui
}
\description{
Script that defines the user interface of Dynameta shiny app
}
\keyword{datasets}
